Refactor code classes
