+++ data_db.rb
@@ -1,0 +1,17 @@
+module DataDb
+  @books = []
+  @persons = []
+  @rentals = []
+
+  def self.books
+    @books
+  end
+
+  def self.persons
+    @persons
+  end
+
+  def self.rentals
+    @rentals
+  end
+end
